Create a Code of Conduct
I used my email address for the time being and am happy to field COC complaints until we have a formal process for dealing with them internally, but I'd like to circle back to this next week so we can develop one. It shouldn't be that difficult, I just want to make sure we're all on the same page.