Only require onCheckedChange to handle boolean values
This makes using CheckboxItem with TypeScript a bit simpler, as it
doesn't require state managed up the tree to allow indeterminate states
if it doesn't want to. More practically, it makes this allowed:

    const [checked, setChecked] = useState(false);
    <CheckboxItem
      checked={checked}
      onCheckedChange={setChecked}
    />

Previously that would result in a type error, because `checked` didn't
permit indeterminate states. However, CheckboxItem never actually sets
itself into an indeterminate state -- that can only be set via props --
meaning we shouldn't have to support that case if we don't want to.