Don’t scroll menus in response to hover
By default, when a menu item is focused, browsers will scroll it into view.
Sometimes even into the centre of the view. This is good when using a keyboard,
because the focused item is automatically visible, but when it was focused in
response to mouse movement, it has the effect of moving the item out from
underneath the pointer.

Fixes radix-ui/primitives#1566