fix(Popper): use `visibility` to hide instead of `opacity` (#2744)
* fix(Popper): use `visibility` to hide instead of `opacity`

When using `hideWhenDetached`, the opacity is changed to 0 instead of the
visibility. This lets users click items that they cannot see. This is
generally a bad user experience.

This patch updates the `Popper` to set the `visibility` to `hidden` when
`hideWhenDetached` is used. This ensures the user cannot interact with
the hidden element.

This aligns with what is in the `@floating-ui/react-dom` documentation.

Ref: https://floating-ui.com/docs/hide#usage

Closes: https://github.com/radix-ui/primitives/issues/2743

* Update packages/react/popper/src/Popper.tsx

* Create 5793010b.yml

---------

Co-authored-by: Benoît Grélard <benoit.grelard@gmail.com>