feat(mergeProps): handlers should return the value of the child handler
In some cases, we can return values from the event handlers. The best use-case I have for this is the following:

```tsx
<Button
  onClick={() => mutateAsync({...})}
>
```

In a case like this, we can use the fact that `mutateAsync` returns a `Promise` to automatically turn on a "loading" state on our button, and turn it off when the promise resolves/rejects.

But if `Slot` doesn't "forward" the return value of the event child handler, this use-case becomes impossible if Button is in a Slot

```tsx
<Tooltip>
  <Button onClick={() => mutateAsync({...})} />
</Tooltip>
```

This PR proposes that `Slot` should return the return value of the child handler. This makes `Slot` even more "transparent" to developers using it.