feat: add the possibility to set the item key of a select item
to enable more control for the developer over the key property of the
select item. the initial behaviour is setting the item value directly as
key property, when working with values that need to be listed separately
but have the same values (for example countries that have the same
country code, but need to be displayed separately) the implemenation
before created a react error in the console because of duplicated keys
in the select item.

this change is backwards compatible to the initial implementation.