fix(Presence): handle escape characters in animation names correctly (#3466)
Fixes #2763.

Using [CSS.escape](https://developer.mozilla.org/en-US/docs/Web/API/CSS/escape_static).

As I checked, this API is supported in all browsers even in the
[2022 browserslist targets](https://github.com/radix-ui/primitives/discussions/1019#discussioncomment-3359338)
of this project.

Further explanation of the bug:

`getComputedStyle()` returns a `CSSStyleDeclaration`, which is
serialized according to the spec.
Meanwhile, the values from `AnimationEvent` are not required to be
serialized (which makes sense, as it's intended for JavaScript usage,
where escaping for CSS syntax is unnecessary). Therefore the
`animationName` is returned as-is in the event handler, causing the
mismatch.